Moved MVCView to mvc.py
